area of rhombus

+++ area_of_rhombus.cpp
@@ -1,9 +1,7 @@
 #include <iostream>
 
 float area_of_rhombus(float q,float p) { 
-// float result = (q * p) / 2;
-// return q * p / 2;
-return (float) q * p / 2;
+return q * p / 2;
 }
 
 int main()
